refactored: components structure and imports
	изменено:      packages/client/src/components/App.tsx
	переименовано: packages/client/src/components/atoms/Button.tsx -> packages/client/src/components/atoms/Button/Button.tsx
	новый файл:    packages/client/src/components/atoms/Button/index.ts
	переименовано: packages/client/src/components/atoms/ErrorLine.tsx -> packages/client/src/components/atoms/ErrorLine/ErrorLine.tsx
	новый файл:    packages/client/src/components/atoms/ErrorLine/index.ts
	переименовано: packages/client/src/components/atoms/Input.tsx -> packages/client/src/components/atoms/Input/Input.tsx
	новый файл:    packages/client/src/components/atoms/Input/index.ts
	изменено:      packages/client/src/components/atoms/index.ts
	переименовано: packages/client/src/components/molecules/Field.tsx -> packages/client/src/components/molecules/Field/Field.tsx
	новый файл:    packages/client/src/components/molecules/Field/index.ts
	изменено:      packages/client/src/components/molecules/index.ts
	переименовано: packages/client/src/components/organisms/LoginForm.tsx -> packages/client/src/components/organisms/LoginForm/LoginForm.tsx
	новый файл:    packages/client/src/components/organisms/LoginForm/index.ts
	изменено:      packages/client/src/components/organisms/index.ts
	удалено:       packages/client/src/components/pages/Login.tsx
	новый файл:    packages/client/src/components/pages/LoginPage/LoginPage.tsx
	новый файл:    packages/client/src/components/pages/LoginPage/index.ts
	изменено:      packages/client/src/components/pages/index.ts
	изменено:      packages/client/src/main.tsx
