+++ packages/client/src/components/App.tsx
@@ -1,4 +1,4 @@
-import LoginPage from "./pages/Login";
+import { LoginPage } from "@/components";
 
 function App() {
   return <LoginPage />;
