+++ packages/client/src/components/atoms/ErrorLine/index.ts
@@ -1,0 +1,1 @@
+export { default as ErrorLine } from "./ErrorLine";
