+++ packages/client/src/components/organisms/LoginForm/index.ts
@@ -1,0 +1,1 @@
+export { default as LoginForm } from "./LoginForm";
