+++ packages/client/src/components/pages/LoginPage/LoginPage.tsx
@@ -1,0 +1,9 @@
+import { LoginForm } from "@/components";
+
+export default function LoginPage(props: unknown) {
+  return (
+    <main className="h-screen bg-gray-800 flex mx-auto justify-center items-center">
+      <LoginForm />
+    </main>
+  );
+}
