+++ packages/client/src/main.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import App from "./components/App";
+import App from "@/components/App";
 import "./index.css";
 import { Provider } from "react-redux";
 import store from "./store/configure";
